gatewayapi: move listener and route status to separate files
